sql-vuln-detection/terraform/app/requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.10
prometheus_client==0.19.0
